message_reader module
Added a bare-bones module for reading user messages and returning bot responses. Also removed comments in bot module, as the previous plan has changed to one where the message_reader receives a user message (as a string), processes, and returns a bot message (as a string) so it doesn't bloat the bot module. This leads to better readability and keeping processes separate.